Create file with environment variables and aliases
Create a file when the Snap package is installed or refreshed which
exports the JavaFX SDK environment variables.

Replace the 'gradle' plugin with the 'nil' plugin for full control
over the build script and packages.

Build with the 'default-jdk-*' packages instead of 'openjdk-11-*'
so that Javadoc links to externally referenced classes can work even
for OpenJDK 8 on Ubuntu 16.04 LTS.

Add the GPLv2 license header to the 'snapcraft.yaml' file. Rewrite
its summary and description.

Add the Oracle TRADEMARK file for OpenJDK.

Update the README file to reflect the changes.