Upgrade to the core20 and openjdk snap packages
Align the build of this snap package with the automated build of
OpenJFX on GitHub, which use Ubuntu 20.04 and OpenJDK 15.

Switch to the core20 base snap, for which the Gradle plugin is not
available. The Gradle plugin defines the Java proxy system properties
and adds the Ubuntu JDK and JRE packages of OpenJDK 11. Instead of
the plugin, define the proxy system properties in the build scriptlet
and install the OpenJDK 15 snap package for the Java Development Kit.

Extract the source ZIP file from the build scriptlet for access to
the JAVA_HOME environment variable.