Add a custom runtime image for JavaFX applications
Add a custom Java runtime image that includes all of the JavaFX
modules and their dependencies. Define the command 'openjfx.java'
for launching JavaFX applications. The runtime image adds 93 MB to
the size of the Snap package, but it lets you run JavaFX applications
with the latest release of Java even on older systems that don't have
OpenJDK 11 or later.

The 'openjfx.java' command is also more convenient than the normal
Java launcher because it doesn't require the '--module-path' and
'--add-modules' options for the JavaFX modules.

Add the build number to the JavaFX version string and replace
'-internal' with '-snap' as shown in the example below:

  $ cat sdk/lib/javafx.properties
  javafx.version=16-snap
  javafx.runtime.version=16-snap+5-2021-01-02-003446
  javafx.runtime.build=5