Add a content interface and update JDK location
Add a content interface to provide a read-only copy of the JavaFX
SDK and JMOD archives to other Snap packages. Name the interface
'javafx-NN-2004' where NN is the JavaFX release number and 2004
denotes the Ubuntu 20.04 LTS build environment.

Update the location of the JDK to its new location within the OpenJDK
Snap package ('jdk' instead of 'jvm').

Rebuild for a security update to the 'libproxy1v5' package:

  USN-4673-1: libproxy vulnerability
  https://ubuntu.com/security/notices/USN-4673-1