Fix Git workflow and don't build from main branch
Follow the workflow recommended by Junio Hamano, the core maintainer of
Git, for maintaining permanent parallel branches.[1] Put common code
on one branch, create a separate branch for each custom deployment,
and stick to the following two rules:

- You make edits to common files only on the common branch.
- You merge from common to deployment, never the other way.

Make the main branch the branch with the common code and without a
snapcraft.yaml file. Create branches named after the candidate, beta,
and edge channels that contain their specific snapcraft.yaml files.

Don't build from the main branch on GitHub to the edge channel. Create
a new edge branch for the channel and build it on Launchpad. That will
allow updates to the main branch without triggering an unnecessary
build, such as updating the README file or adding an issue template.

[1] Re: Is there a way to exclude user-specified files or directories
from participating in merges?
https://www.spinics.net/linux/lists/git/msg94767.html