Fix "LC_ALL: cannot change locale" on edge
When the LC_ALL environment variable is defined with a language code,
the 'openjfx' command prints the following warning message:

  $ LC_ALL=en_CA.UTF-8 openjfx
  /bin/bash: warning:
    setlocale: LC_ALL: cannot change locale (en_CA.UTF-8)
  /snap/openjfx/19/bin/openjfx.env

The only locale names present in the core18 base are C.UTF-8 along with
C and POSIX, which use the US-ASCII encoding (ANSI_X3.4-1968). Remove
the warning by overriding LC_ALL with C.UTF-8 for the command.

The problem occurs on the Raspberry Pi Desktop, which defines the
LC_ALL environment variable system wide, as shown below, when you
set the locale with the Raspberry Pi Configuration program.

  $ cat /etc/default/locale
  LANG=en_CA.UTF-8
  LANGUAGE=en_CA.UTF-8
  LC_ALL=en_CA.UTF-8