Upgrade to core22 base and remove 32-bit builds
Upgrade to the core22 base for gcc version 11. Building with earlier
versions of gcc causes a runtime error like the following:

  Exception in thread "main" java.lang.reflect.InvocationTargetException
  Caused by: java.lang.UnsatisfiedLinkError: libglassgtk3.so:
  undefined symbol: _ZN15PlatformSupport17OBSERVED_SETTINGSE

Remove the builds for the 32-bit armhf and i386 architectures. Building
for armhf or i386 fails with the error:

  modules/javafx.graphics/src/main/native-glass/gtk/screencast_portal.h:38:31:
  error: format ‘%lu’ expects argument of type ‘long unsigned int’,
  but argument 3 has type ‘uint64_t {aka long long unsigned int}’
  [-Werror=format=]

Remove all of the library packages staged from the core22 base. Provide
only the native libraries that are part of the JavaFX SDK itself.